Refactored to use a mask layer system
You can now use a txt with numbers, which shows which layer each cell is on. When checking the validity of the cell, it will check if the cell is on layer -1, which counts as invalid. If you use the mask 0, it won't matter which layer the cells are on. For layers 1-9, it will only find/return/etc with other cells on it's given layer.